swagger in interface, code in class, life is good
