+++ src/main/java/org/icgc/argo/workflow_management/controller/RunApi.java
@@ -4,29 +4,16 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import javax.validation.Valid;
-import lombok.val;
 import org.icgc.argo.workflow_management.controller.model.RunRequest;
 import org.icgc.argo.workflow_management.controller.model.RunResponse;
 import org.icgc.argo.workflow_management.exception.model.ErrorResponse;
-import org.icgc.argo.workflow_management.service.WorkflowExecutionService;
-import org.icgc.argo.workflow_management.service.model.WESRunParams;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
-@RestController
-@RequestMapping("/runs")
-@Api(value = "WorkflowExecutionService", tags = "WorkflowExecutionService")
-public class RunController {
+import javax.validation.Valid;
 
-  @Autowired
-  @Qualifier("nextflow")
-  private WorkflowExecutionService nextflowService;
+@Api(value = "WorkflowExecutionService", tags = "WorkflowExecutionService")
+public interface RunApi {
 
   @ApiOperation(
       value = "Run a workflow",
@@ -63,23 +50,5 @@ public class RunController {
             message = "An unexpected error occurred.",
             response = ErrorResponse.class)
       })
-  @PostMapping
-  private Mono<RunResponse> postRun(@Valid @RequestBody RunRequest runRequest) {
-
-    val wesService = resolveWesType("nextflow");
-
-    // create run config from request
-    val runConfig =
-        WESRunParams.builder()
-            .workflowUrl(runRequest.getWorkflowUrl())
-            .workflowParams(runRequest.getWorkflowParams())
-            .workflowEngineParameters(runRequest.getWorkflowEngineParameters())
-            .build();
-
-    return wesService.run(runConfig);
-  }
-
-  private WorkflowExecutionService resolveWesType(String workflowType) {
-    return nextflowService;
-  }
+  Mono<RunResponse> postRun(@Valid @RequestBody RunRequest runRequest);
 }
